Added cookies to log voting
Instead of a simple JavaScript hack (saving a data element in the DOM)
to record which quotes a user has already voted for, this data is now
stored in a cookie, so it persists even if they reload the page. This
should make repeat voting a little harder. The cookies are handled with
the Cookies.js library, which is fairly lightweight.