Upgraded to Django 1.11 and added django-cron
The quotes server was upgraded to Django 1.11, so I made some minor
changes to fix things that broke with the upgrade (e.g. the syntax of
urls.py changed). Also, I used the django-cron module to add a cron-like
task that regularly checks for new unapproved quotes and e-mails the
site administrator a reminder.

Note that django-cron only works because I added a Linux cron task to my
user's crontab on the quotes server that executes "manage.py runcrons"
every 5 minutes.