Added a new "random quote" page
I realized the random-quote JSON API was fun but wasn't getting used, so
I made a simple HTML page that does the same thing. Right now there's no
explicit link to use the filter-by-year function, but it works if you
enter the year in the URL. Maybe I'll add more features in the future.