Fixed JavaScript loading order
Standard practice is now to put JavaScript at the bottom of the <body> tag rather than inside the <head>, so that it doesn't load until after the page finishes loading. Also, updated the CSRF cookie setup script to wait until the document is ready to retrieve the cookie.

This should fix a bug I noticed where the upvote buttons sometimes don't work because the ajax-setup script ran before the CSRF cookie was set, and thus AJAX requests don't have a CSRF cookie to send.